refactor(anime): extract AnimeCarousel from the page
Move the carousel state and navigation handlers into their own
component so anime.vue is layout-only.
